fix(db): cast to numeric for ROUND in calibration SQL
Postgres has no round(double precision, int); pi()/power()/acos()/sqrt() all
return double precision, so ROUND(expr, 2) failed at function creation. Cast
the rounded expressions to numeric in calib_volume/calib_tolerance and the
charts capacity query.

+++ backend/src/routes/calibration.js
@@ -10,7 +10,7 @@ router.get('/charts', authenticate, async (_req, res, next) => {
   try {
     const { rows } = await pool.query(
       `SELECT id, name, diameter_cm, length_cm,
-              ROUND(pi() * power(diameter_cm/2, 2) * length_cm / 1000) AS capacity_ltrs
+              ROUND((pi() * power(diameter_cm/2, 2) * length_cm / 1000)::numeric) AS capacity_ltrs
        FROM tank_calibration_charts
        WHERE is_active
        ORDER BY diameter_cm, length_cm`
